node-fetch moved to es_module, server.js now uses imports not requires
